Fix Gaussian summary NAs, support double observations, add EIP plot
- Fix summary() param names for Gaussian: was using n_colors/2 instead
  of n_colors, producing NAs for mu_2 and sigma2_2
- Change Observations storage from int to double so Gaussian data no
  longer needs rounding to integers
- Update EmissionLikelihood/LogLikelihood/UpdateEmissionParams to iterate
  over double observations
- Add "edge_inclusion" option to plot() method showing graph with edge
  widths scaled by inclusion probability (requires igraph)
- Increase vignette MCMC iterations to 5000 for better trace plots
- Add edge inclusion probability plots to emission models vignette

+++ inst/include/mdgm/emission.hpp
@@ -11,7 +11,7 @@ namespace mdgm {
 
 // Returns likelihood p(y_i | z_i = k, theta) for each color k
 std::vector<double> EmissionLikelihood(
-    std::span<const int> y_i, std::span<const double> theta,
+    std::span<const double> y_i, std::span<const double> theta,
     std::size_t ncolors, FamilyType type);
 
 // Log-likelihood of all observations: sum_i log p(y_i | z_i, theta)
